Message bus adapter using RabbitMQ
Added the module Exchange.Adapter.RabbitBus which as the behaviour Exchange.MessageBus. This module uses the modules Exchange.Adapter.RabbitBus.Producer
and Exchange.Adapter.RabbitBus.Consumer. The first is a server that receives messages ({:cast_event, event, payload}) and publishes a encoded message to a RabbitMQ exchange. The second is also a server that maintains a map of event => list(subscribers) and it consumes messages from RabbitMQ, these messages are decoded and then broadcasted to the respective subscribers.

Added Jason encode protocol to
-OrderCancelled
-OrderExpired
-OrderPlaced
-OrderQueued
-PriceBroadcast
-TradeExecuted
-TradeProcessed(Necessary to test decode and encode on this)
-Trade
-Order

Corrected the setup_time_series/0 and setup_message_bus/0

The function validate_config/1 was changed to validate_config/0 in the behaviour module Exchange.TimeSeries and Exchange.MessageBus. The configuration is retrieved by calling Application.get_env/3.

Removed the module Exchange.Adapter.Flux.Connection and the module Exchange.Adapter.Flux manages the connection.